move mocks into the __mocks__ folder
