+++ public/app/plugins/datasource/influxdb/__mocks__/response.ts
@@ -1,108 +1,5 @@
-import { of } from 'rxjs';
-
-import {
-  AdHocVariableFilter,
-  DataQueryRequest,
-  DataSourceInstanceSettings,
-  dateTime,
-  FieldType,
-  PluginType,
-  ScopedVars,
-} from '@grafana/data';
-import {
-  BackendDataSourceResponse,
-  FetchResponse,
-  getBackendSrv,
-  setBackendSrv,
-  VariableInterpolation,
-} from '@grafana/runtime';
-import { SQLQuery } from '@grafana/sql';
-
-import { TemplateSrv } from '../../../features/templating/template_srv';
-
-import InfluxDatasource from './datasource';
-import { InfluxOptions, InfluxQuery, InfluxVersion } from './types';
-
-const getAdhocFiltersMock = jest.fn().mockImplementation(() => []);
-const replaceMock = jest.fn().mockImplementation((a: string, ...rest: unknown[]) => a);
-
-export const templateSrvStub = {
-  getAdhocFilters: getAdhocFiltersMock,
-  replace: replaceMock,
-} as unknown as TemplateSrv;
-
-export function mockTemplateSrv(
-  getAdhocFiltersMock: (datasourceName: string) => AdHocVariableFilter[],
-  replaceMock: (
-    target?: string,
-    scopedVars?: ScopedVars,
-    format?: string | Function | undefined,
-    interpolations?: VariableInterpolation[]
-  ) => string
-): TemplateSrv {
-  return {
-    getAdhocFilters: getAdhocFiltersMock,
-    replace: replaceMock,
-  } as unknown as TemplateSrv;
-}
-
-export function mockBackendService(response: FetchResponse) {
-  const fetchMock = jest.fn().mockReturnValue(of(response));
-  const origBackendSrv = getBackendSrv();
-  setBackendSrv({
-    ...origBackendSrv,
-    fetch: fetchMock,
-  });
-  return fetchMock;
-}
-
-export function getMockInfluxDS(
-  instanceSettings: DataSourceInstanceSettings<InfluxOptions> = getMockDSInstanceSettings(),
-  templateSrv: TemplateSrv = templateSrvStub
-): InfluxDatasource {
-  return new InfluxDatasource(instanceSettings, templateSrv);
-}
-
-export function getMockDSInstanceSettings(
-  overrideJsonData?: Partial<InfluxOptions>
-): DataSourceInstanceSettings<InfluxOptions> {
-  return {
-    id: 123,
-    url: 'proxied',
-    access: 'proxy',
-    name: 'influxDb',
-    readOnly: false,
-    uid: 'influxdb-test',
-    type: 'influxdb',
-    meta: {
-      id: 'influxdb-meta',
-      type: PluginType.datasource,
-      name: 'influxdb-test',
-      info: {
-        author: {
-          name: 'observability-metrics',
-        },
-        version: 'v0.0.1',
-        description: 'test',
-        links: [],
-        logos: {
-          large: '',
-          small: '',
-        },
-        updated: '',
-        screenshots: [],
-      },
-      module: '',
-      baseUrl: '',
-    },
-    jsonData: {
-      version: InfluxVersion.InfluxQL,
-      httpMode: 'POST',
-      dbName: 'site',
-      ...(overrideJsonData ? overrideJsonData : {}),
-    },
-  };
-}
+import { FieldType } from '@grafana/data';
+import { BackendDataSourceResponse, FetchResponse } from '@grafana/runtime';
 
 export const mockInfluxFetchResponse = (
   overrides?: Partial<FetchResponse<BackendDataSourceResponse>>
@@ -229,101 +126,6 @@ export const mockInfluxRetentionPolicyResponse = [
     data: { values: [['autogen', 'bar', '5m_avg', '1m_avg', 'default']] },
   },
 ];
-
-type QueryType = InfluxQuery & SQLQuery;
-
-export const mockInfluxQueryRequest = (targets?: QueryType[]): DataQueryRequest<QueryType> => {
-  return {
-    app: 'explore',
-    interval: '1m',
-    intervalMs: 60000,
-    range: {
-      from: dateTime(0),
-      to: dateTime(10),
-      raw: { from: dateTime(0), to: dateTime(10) },
-    },
-    rangeRaw: {
-      from: dateTime(0),
-      to: dateTime(10),
-    },
-    requestId: '',
-    scopedVars: {},
-    startTime: 0,
-    targets: targets ?? mockTargets(),
-    timezone: '',
-  };
-};
-
-export const mockTargets = (): QueryType[] => {
-  return [
-    {
-      refId: 'A',
-      datasource: {
-        type: 'influxdb',
-        uid: 'vA4bkHenk',
-      },
-      policy: 'default',
-      resultFormat: 'time_series',
-      orderByTime: 'ASC',
-      tags: [],
-      groupBy: [
-        {
-          type: 'time',
-          params: ['$__interval'],
-        },
-        {
-          type: 'fill',
-          params: ['null'],
-        },
-      ],
-      select: [
-        [
-          {
-            type: 'field',
-            params: ['value'],
-          },
-          {
-            type: 'mean',
-            params: [],
-          },
-        ],
-      ],
-      measurement: 'cpu',
-    },
-  ];
-};
-
-export const mockInfluxQueryWithTemplateVars = (adhocFilters: AdHocVariableFilter[]): InfluxQuery => ({
-  refId: 'x',
-  alias: '$interpolationVar',
-  measurement: '$interpolationVar',
-  policy: '$interpolationVar',
-  limit: '$interpolationVar',
-  slimit: '$interpolationVar',
-  tz: '$interpolationVar',
-  tags: [
-    {
-      key: 'cpu',
-      operator: '=~',
-      value: '/^$interpolationVar,$interpolationVar2$/',
-    },
-  ],
-  groupBy: [
-    {
-      params: ['$interpolationVar'],
-      type: 'tag',
-    },
-  ],
-  select: [
-    [
-      {
-        params: ['$interpolationVar'],
-        type: 'field',
-      },
-    ],
-  ],
-  adhocFilters,
-});
 
 export const mockInfluxSQLFetchResponse: FetchResponse<BackendDataSourceResponse> = {
   config: {
