+++ public/app/plugins/datasource/influxdb/influxql_metadata_query.test.ts
@@ -1,7 +1,7 @@
 import config from 'app/core/config';
 
+import { getMockInfluxDS } from './__mocks__/datasource';
 import { getAllMeasurements, getAllPolicies, getFieldKeys, getTagKeys, getTagValues } from './influxql_metadata_query';
-import { getMockInfluxDS } from './mocks';
 import { InfluxQuery } from './types';
 
 describe('influx_metadata_query', () => {
