+++ public/app/plugins/datasource/influxdb/influxql_query_builder.test.ts
@@ -1,5 +1,5 @@
+import { templateSrvStub as templateService } from './__mocks__/datasource';
 import { buildMetadataQuery } from './influxql_query_builder';
-import { templateSrvStub as templateService } from './mocks';
 import { DEFAULT_POLICY } from './types';
 
 describe('influxql-query-builder', () => {
