+++ public/app/plugins/datasource/influxdb/response_parser.test.ts
@@ -6,8 +6,8 @@ import { FetchResponse } from '@grafana/runtime';
 import config from 'app/core/config';
 import { backendSrv } from 'app/core/services/backend_srv'; // will use the version in __mocks__
 
+import { getMockDSInstanceSettings, getMockInfluxDS } from './__mocks__/datasource';
 import InfluxQueryModel from './influx_query_model';
-import { getMockDSInstanceSettings, getMockInfluxDS } from './mocks';
 import ResponseParser, { getSelectedParams } from './response_parser';
 import { InfluxQuery } from './types';
